_get_XXXX_lock() moved to utils.py
